requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
pdfminer.six
cachetools
mangum
